Configure expect().toIncludeQueryParam() to be available in all tests
